[#50] 공연장 공연 홀 병합 기능 수정
* 예외 처리
    1. 공연장 수정 시 존재하지 않는 id로 업데이트할 경우 VenueIdNotExistsException 예외 처리
    2 .공연 홀 수정 시 존재하지 않는 id로 업데이트할 경우 VenueHallIdNotExistsException 예외 처리
    3. 공연 홀 삭제 시 존재하지 않는 id로 삭제할 경우 VenueHallIdNotExistsException 예외 처리
* VenueUpdateRequest DTO에 updateHallIds 와 VenueHallRequest를 VenueHallUpdateRequest로 통합
   =>만약 클라이언트에서 updateHallIds 속성 1개를 보냈는데 VenueHallUpdateRequest에서는 속성 2개를 보낼 시 서버에서 로직을 처리해 줘야 함
     따라서 통합하는 게 더 가독성이 좋아 보여 통합
* 수정, 삭제할 공연 홀 id가 DB에 존재하는지 로직 처리
    1. VenueHallService의 checkVenueHallIdsExists로 수정,삭제할 공연 홀 id들이 DB에 존재하는지 로직 처리
    2. VenueHallMapper의 getVenueHallCount로 업데이트할 공연 홀 id들이 DB에 몇개 존재하는지 반환 시켜줌
    3. 총 갯수가 아니라면 예외 처리
* 공연장 공연 홀 테스트 기능 추가

+++ src/main/java/com/show/showticketingservice/service/VenueService.java
@@ -29,21 +29,28 @@ public class VenueService {
     }
 
     public void checkVenueExists(String venueName) {
-        if (venueMapper.isVenueExists(venueName)) {
+        if(venueMapper.isVenueExists(venueName)) {
             throw new VenueAlreadyExistsException();
         }
     }
 
     @Transactional
     public void updateVenueInfo(int venueId, VenueUpdateRequest venueUpdateRequest) {
+
+        checkVenueIdExists(venueId);
+
         if(venueUpdateRequest.getVenue() != null) {
             checkVenueExists(venueUpdateRequest.getVenue().getName());
             venueMapper.updateVenueInfo(venueId, venueUpdateRequest.getVenue());
         }
 
-        venueHallService.updateVenueHalls(venueUpdateRequest.getUpdateHallIds(), venueUpdateRequest.getVenueHallRequests(), venueId);
+        if(!venueUpdateRequest.getVenueHallUpdateRequests().isEmpty()) {
+            venueHallService.updateVenueHalls(venueId, venueUpdateRequest.getVenueHallUpdateRequests());
+        }
 
-        venueHallService.deleteVenueHalls(venueId, venueUpdateRequest.getDeleteHallIds());
+        if(!venueUpdateRequest.getDeleteHallIds().isEmpty()) {
+            venueHallService.deleteVenueHalls(venueId, venueUpdateRequest.getDeleteHallIds());
+        }
     }
 
     public void deleteVenue(int venueId) {
